requirements.txt
requests
psycopg[binary]
python-dotenv